add comments route working
