+++ controllers/home-routes.js
@@ -1,6 +1,6 @@
 const router = require('express').Router();
-const { Category, Recipe } = require('../models');
-// const withAuth = require('../utils/auth');
+const { Category, Recipe, Comment, User } = require('../models');
+const withAuth = require('../utils/auth');
 
 router.get('/login', async (req, res) => {
   // If the user is already logged in, redirect the request to another route
@@ -54,23 +54,33 @@ router.get('/:id', async (req, res) => {
 });
 
 // get recipe by id
-// router.get('/recipes/:id', withAuth, async (req, res) => {
-//   try {
-//     const loggedIn = req.session.logged_in;
-//     const categoryData = await Category.findAll();
-//     const categories = categoryData.map((category) =>
-//       category.get({ plain: true })
-//     );
-//     const recipeData = await Recipe.findByPk(req.params.id, {
-//       where: {
-//         user_id: req.session.user_id,
-//       },
-//     });
-//     const recipe = recipeData.get({ plain: true });
-//     res.render('recipes', { recipe, categories, logged_in: loggedIn });
-//   } catch (err) {
-//     res.status(500).json(err);
-//   }
-// });
+router.get('/recipes/:id',withAuth, async (req, res) => {
+  try {
+    const categoryData = await Category.findAll();
+    const categories = categoryData.map((category) =>
+      category.get({ plain: true })
+    );
+    const recipeData = await Recipe.findByPk(req.params.id, {
+      include: [
+        {
+          model: Comment,
+          include: {
+            model: User,
+          },
+        },
+      ],
+    });
+    const recipe = recipeData.get({ plain: true });
+    res.render('recipes', {
+      recipe,
+      categories,
+      logged_in: req.session.logged_in,
+      login: true,
+    });
+    // res.json(recipeData)
+  } catch (err) {
+    res.status(500).json(err);
+  }
+});
 
 module.exports = router;
