+++ views/all.handlebars
@@ -1,7 +1,7 @@
     {{! recipe by category id }}
     {{#each recipePK.recipes as |recipe|}}
         <div class='recipe-content'>
-      <a href='/api/recipes/{{recipe.id}}'>
+      <a href='/recipes/{{recipe.id}}'>
           <img
             src='/assets/uploads/{{recipe.photo}}'
             class='rounded py-9 px-9 d-flex align-content-center'
@@ -18,7 +18,7 @@
     {{! all recipes }}
     {{#each recipes as |recipe|}}
       <div class='recipe-content'>
-        <a href='/api/recipes/{{recipe.id}}'>
+        <a href='/recipes/{{recipe.id}}'>
           <img
             src='/assets/uploads/{{recipe.photo}}'
             class='rounded py-9 px-9 d-flex align-content-center'
