+++ views/profile.handlebars
@@ -12,7 +12,7 @@
       {{#each recipes as |recipe|}}
         <div class='row mb-2'>
           <div class='col-md-8'>
-            <a href='/api/recipes/{{recipe.id}}'><h4>{{name}}</h4>
+            <a href='/recipes/{{recipe.id}}'><h4>{{name}}</h4>
               <img
                 src='/assets/uploads/{{photo}}'
                 class='rounded py-9 px-9 d-flex align-content-center'
@@ -42,7 +42,7 @@
      
         <div class='row mb-2'>
           <div class='col-md-8'>
-            <a href='/api/recipes/{{id}}'><h4>{{name}}</h4>
+            <a href='/recipes/{{id}}'><h4>{{name}}</h4>
               <img
                 src='/assets/uploads/{{photo}}'
                 class='rounded py-9 px-9 d-flex align-content-center'
@@ -50,7 +50,7 @@
                 width='500px'
                 alt='{{name}}'
               /></a>
-            <a href='/api/recipes/{{id}}'>{{recipe.name}}</a>
+            <a href='/recipes/{{id}}'>{{recipe.name}}</a>
           </div>
           <div class='col-md-4'>
             <button class='delBut btn btn-sm btn-danger' user-id='{{id}}'><i
